# openapi.yaml
openapi: 3.0.3
info:
  title: URL Shortener API
  version: 1.0.0
servers:
  - url: http://localhost:8080
components:
  schemas:
    CreateLinkRequest:
      type: object
      required: [targetUrl]
      properties:
        targetUrl: { type: string, maxLength: 2048, example: "https://example.com" }
        customAlias: { type: string, minLength: 3, maxLength: 32 }
        expiresAt: { type: string, format: date-time, nullable: true }
    LinkResponse:
      type: object
      properties:
        shortCode: { type: string }
        shortUrl: { type: string, format: uri }
        targetUrl: { type: string, format: uri }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
    Breakdown:
      type: object
      properties:
        value: { type: string }
        count: { type: integer, format: int64 }
    AnalyticsResponse:
      type: object
      properties:
        shortCode: { type: string }
        totalClicks: { type: integer, format: int64 }
        uniqueVisitors: { type: integer, format: int64 }
        clicksLast24h: { type: integer, format: int64 }
        firstClickAt: { type: string, format: date-time, nullable: true }
        lastClickAt: { type: string, format: date-time, nullable: true }
        topReferrers: { type: array, items: { $ref: '#/components/schemas/Breakdown' } }
        devices: { type: array, items: { $ref: '#/components/schemas/Breakdown' } }
        asOf: { type: string, format: date-time }
    Problem:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
  responses:
    Problem:
      description: RFC 7807 error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
paths:
  /api/v1/links:
    post:
      summary: Create a short link
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateLinkRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LinkResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
        '503': { $ref: '#/components/responses/Problem' }
  /api/v1/links/{shortCode}:
    parameters:
      - name: shortCode
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Get link metadata
      responses:
        '200':
          description: Link metadata
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LinkResponse' }
        '404': { $ref: '#/components/responses/Problem' }
    delete:
      summary: Soft-delete a link
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/Problem' }
  /api/v1/links/{shortCode}/stats:
    get:
      summary: Get click analytics
      parameters:
        - name: shortCode
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Analytics snapshot
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnalyticsResponse' }
        '404': { $ref: '#/components/responses/Problem' }
  /{shortCode}:
    get:
      summary: Redirect to the target URL
      parameters:
        - name: shortCode
          in: path
          required: true
          schema: { type: string, pattern: '^[a-zA-Z0-9_-]{3,32}$' }
      responses:
        '302': { description: Redirect, headers: { Location: { schema: { type: string, format: uri } } } }
        '404': { $ref: '#/components/responses/Problem' }
        '410': { $ref: '#/components/responses/Problem' }
        '503': { $ref: '#/components/responses/Problem' }
  /healthz:
    get:
      summary: Liveness
      responses:
        '200': { description: Process is live }
  /readyz:
    get:
      summary: Dependency readiness
      responses:
        '200': { description: Ready }
        '503': { description: Not ready }